adding blueprints and code reconstruction complete
