+++ app/auth/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp = Blueprint('auth', __name__)
+
+
+from app.auth import routes
