+++ app/auth/forms.py
@@ -1,8 +1,6 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, BooleanField, SubmitField
-from wtforms import TextAreaField
 from wtforms.validators import DataRequired, ValidationError, Email, EqualTo
-from wtforms.validators import Optional, Length
 from app.models import User
 
 
@@ -34,20 +32,6 @@ class RegistrationForm(FlaskForm):
         user = User.query.filter_by(email=email.data).first()
         if user is not None:
             raise ValidationError('Please use a different email address.')
-
-
-class PostForm(FlaskForm):
-    clientname = StringField('Name', validators=[DataRequired()])
-    clientss = StringField('Social Security number', validators=[Optional()])
-    clientemail = StringField('Email', validators=[DataRequired()])
-    clientphone = StringField('Phone', validators=[DataRequired()])
-    clientaddress = StringField('Address', validators=[Optional()])
-    clientzip = StringField('ZIP', validators=[Optional()])
-    clientcity = StringField('City', validators=[Optional()])
-    clientinfo = TextAreaField('Info',
-                               validators=[Optional(), Length(max=2048)])
-    submit = SubmitField('Submit')
-    cancel = SubmitField('Cancel')
 
 
 class ResetPasswordRequestForm(FlaskForm):
