+++ app/email.py
@@ -1,6 +1,6 @@
 from flask_mail import Message
-from app import app, mail
-from flask import render_template
+from app import mail
+from flask import current_app
 from threading import Thread
 
 
@@ -13,15 +13,5 @@ def send_email(subject, sender, recipients, text_body, html_body):
     msg = Message(subject, sender=sender, recipients=recipients)
     msg.body = text_body
     msg.html = html_body
-    Thread(target=send_async_email, args=(app, msg)).start()
-
-
-def send_password_reset_email(user):
-    token = user.get_reset_password_token()
-    send_email('[Spjald] - Reset your password',
-               sender=app.config['ADMINS'][0],
-               recipients=[user.email],
-               text_body=render_template('email/reset_password.txt',
-                                         user=user, token=token),
-               html_body=render_template('email/reset_password.html',
-                                         user=user, token=token))
+    Thread(target=send_async_email,
+           args=(current_app._get_current_object(), msg)).start()
