+++ app/errors/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp = Blueprint('errors', __name__)
+
+
+from app.errors import handlers
