+++ app/main/routes.py
@@ -1,0 +1,124 @@
+from app import db
+from app.main import bp
+from app.main.forms import PostForm
+from app.models import User, Post
+from flask import request, render_template, flash, redirect, \
+    url_for, current_app
+from flask_login import current_user, login_required
+
+
+# Routes
+@bp.route('/', methods=['GET', 'POST'])
+@bp.route('/index', methods=['GET', 'POST'])
+@login_required
+def index():
+    page = request.args.get('page', 1, type=int)
+    posts = current_user.followed_posts().paginate(
+        page, current_app.config['POSTS_PER_PAGE'], False)
+    next_url = url_for('main.index', page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('main.index', page=posts.prev_num) \
+        if posts.has_prev else None
+    return render_template('index.html', title='Home',
+                           posts=posts.items, next_url=next_url,
+                           prev_url=prev_url)
+
+
+@bp.route('/addpost', methods=['GET', 'POST'])
+@login_required
+def addpost():
+    form = PostForm()
+    if form.validate_on_submit():
+        if form.submit.data:
+            post = Post(clientname=form.clientname.data,
+                        clientss=form.clientss.data,
+                        clientemail=form.clientemail.data,
+                        clientphone=form.clientphone.data,
+                        clientaddress=form.clientaddress.data,
+                        clientzip=form.clientzip.data,
+                        clientcity=form.clientcity.data,
+                        clientinfo=form.clientinfo.data,
+                        author=current_user)
+            db.session.add(post)
+            db.session.commit()
+            flash('Client data successfully added!')
+            return redirect(url_for('main.addpost'))
+        else:
+            return redirect(url_for('main.index'))
+    return render_template('addpost.html', title='Add Post', form=form)
+
+
+@bp.route('/editpost/<int:clientid>', methods=['GET', 'POST'])
+@login_required
+def editpost(clientid):
+    qry = Post.query.filter_by(clientid=clientid).first()
+    form = PostForm(request.form, obj=qry)
+    if form.validate_on_submit():
+        if form.submit.data:
+            form.populate_obj(qry)
+            db.session.commit()
+            flash('Your changes have been saved.')
+            return redirect(url_for('main.index', clientid=clientid))
+        else:
+            return redirect(url_for('main.index'))
+    return render_template('editpost.html', title='Edit post',
+                           form=form, clientid=clientid)
+
+
+@bp.route('/deletepost/<int:clientid>', methods=['GET', 'POST'])
+@login_required
+def deletepost(clientid):
+    qry = Post.query.filter_by(clientid=clientid).first()
+    db.session.delete(qry)
+    db.session.commit()
+    flash('Post successfully deleted!')
+    return redirect(url_for('main.index'))
+
+
+@bp.route('/user/<username>')
+@login_required
+def user(username):
+    user = User.query.filter_by(username=username).first_or_404()
+    page = request.args.get('page', 1, type=int)
+    posts = user.posts.order_by(Post.timestamp.desc()).paginate(
+        page, current_app.config['POSTS_PER_PAGE'], False)
+    next_url = url_for('main.user', username=user.username,
+                       page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('main.user', username=user.username,
+                       page=posts.prev_num) \
+        if posts.has_prev else None
+    return render_template('user.html', user=user, posts=posts.items,
+                           next_url=next_url, prev_url=prev_url)
+
+
+@bp.route('/follow/<username>')
+@login_required
+def follow(username):
+    user = User.query.filter_by(username=username).first()
+    if user is None:
+        flash('User %(username)s not found.', username=username)
+        return redirect(url_for('main.index'))
+    if user == current_user:
+        flash(('You cannot follow yourself!'))
+        return redirect(url_for('main.user', username=username))
+    current_user.follow(user)
+    db.session.commit()
+    flash('You are following %(username)s!', username=username)
+    return redirect(url_for('main.user', username=username))
+
+
+@bp.route('/unfollow/<username>')
+@login_required
+def unfollow(username):
+    user = User.query.filter_by(username=username).first()
+    if user is None:
+        flash('User %(username)s not found.', username=username)
+        return redirect(url_for('main.index'))
+    if user == current_user:
+        flash('You cannot unfollow yourself!')
+        return redirect(url_for('main.user', username=username))
+    current_user.unfollow(user)
+    db.session.commit()
+    flash('You are not following %(username)s.', username=username)
+    return redirect(url_for('main.user', username=username))
