+++ app/models.py
@@ -1,7 +1,8 @@
 import jwt
 from datetime import datetime
-from app import app, db, loginm
+from app import db, loginm
 from werkzeug.security import generate_password_hash, check_password_hash
+from flask import current_app
 from flask_login import UserMixin
 from hashlib import _hashlib
 from time import time
@@ -13,11 +14,11 @@ def load_user(id):  # pylint: disable=redefined-builtin
 
 
 # Database models
-followers = db.Table('followers',
-                     db.Column('follower_id', db.Integer,
-                               db.ForeignKey('user.id')),
-                     db.Column('followed_id', db.Integer,
-                               db.ForeignKey('user.id')))
+followers = db.Table(
+    'followers',
+    db.Column('follower_id', db.Integer, db.ForeignKey('user.id')),
+    db.Column('followed_id', db.Integer, db.ForeignKey('user.id'))
+)
 
 
 class User(UserMixin, db.Model):
@@ -70,17 +71,17 @@ class User(UserMixin, db.Model):
     def get_reset_password_token(self, expires_in=600):
         return jwt.encode(
             {'reset_password': self.id, 'exp': time() + expires_in},
-            app.config['SECRET_KEY'], algorithm='HS256').decode('utf-8')
+            current_app.config['SECRET_KEY'],
+            algorithm='HS256').decode('utf-8')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
             id = jwt.decode(  # pylint: disable=redefined-builtin
-                token, app.config['SECRET_KEY'],
+                token, current_app.config['SECRET_KEY'],
                 algorithms=['HS256'])['reset_password']
         except Exception as e:
-            print("EXCEPTION FORMAT PRINT:\n{}".format(e))
-            return
+            return print("EXCEPTION FORMAT PRINT:\n{}".format(e))
         return User.query.get(id)
