--- app/templates/404.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-
-{% block app_content %}
-    <h3>File Not Found</h3>
-    <p><a href="{{ url_for('index') }}">Back</a></p>
-{% endblock %}
