+++ app/templates/_post.html
@@ -13,7 +13,7 @@
                     {{ post.clientaddress }}<br />
                     {{ post.clientzip }}, {{ post.clientcity }}<br /><br />
                     <p>
-                        <a href="{{ url_for('editpost', clientid=post.clientid) }}"><i class="fas fa-edit"></i></a>&nbsp;&nbsp;&nbsp;<a href="{{ url_for('deletepost', clientid=post.clientid) }}"><i class="fas fa-trash-alt"></i></a>
+                        <a href="{{ url_for('main.editpost', clientid=post.clientid) }}"><i class="fas fa-edit"></i></a>&nbsp;&nbsp;&nbsp;<a href="{{ url_for('main.deletepost', clientid=post.clientid) }}"><i class="fas fa-trash-alt"></i></a>
                     </p>
                 </div>
             </td>
