+++ app/templates/auth/login.html
@@ -1,9 +1,7 @@
 {% extends "base.html" %}
-
 {% import 'bootstrap/wtf.html' as wtf %}
-
 {% block app_content %}
-    <h3>Sign In</h3>
+    <h3>Sign in</h3>
     <div class="row">
         <div class="col-md-4">
             {{ wtf.quick_form(form) }}
@@ -11,11 +9,11 @@
     </div>
     <br />
     <p>
-        New User?
-        <a href="{{ url_for('register') }}">Click to Register!</a>
+        New user?
+        <a href="{{ url_for('auth.register') }}">Click to register</a>
     </p>
     <p>
-        Forgot Your Password?
-        <a href="{{ url_for('reset_password_request') }}">Click to Reset It</a>
+        Forgot your password?
+        <a href="{{ url_for('auth.reset_password_request') }}">Click to reset it</a>
     </p>
 {% endblock %}
