+++ app/templates/base.html
@@ -1,9 +1,7 @@
 {% extends 'bootstrap/base.html' %}
-
 {% block title %}
     {% if title %}{{ title }} - Spjald{% else %}Welcome to Spjald{% endif %}
 {% endblock %}
-
 {% block navbar %}
 <script src="https://kit.fontawesome.com/81c190df67.js" crossorigin="anonymous"></script>
     <nav class="navbar navbar-default">
@@ -15,28 +13,27 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a class="navbar-brand" href="{{ url_for('index') }}"><i class="far fa-address-card fa-lg"></i> Spjald</a>
+                <a class="navbar-brand" href="{{ url_for('main.index') }}"><i class="far fa-address-card fa-lg"></i> Spjald</a>
             </div>
             <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
                 <ul class="nav navbar-nav">
-                    <li><a href="{{ url_for('index') }}"><i class="fas fa-home"></i> Home</a></li>
+                    <li><a href="{{ url_for('main.index') }}"><i class="fas fa-home"></i> Home</a></li>
                     {% if current_user.is_authenticated %}
-                    <li><a href="{{ url_for('addpost') }}"><i class="fas fa-plus"></i> Add post</a></li>
+                    <li><a href="{{ url_for('main.addpost') }}"><i class="fas fa-plus"></i> Add post</a></li>
                     {% endif %}
                 </ul>
                 <ul class="nav navbar-nav navbar-right">
                     {% if current_user.is_anonymous %}
-                    <li><a href="{{ url_for('login') }}"><i class="fas fa-sign-in-alt"></i> Login</a></li>
+                    <li><a href="{{ url_for('auth.login') }}"><i class="fas fa-sign-in-alt"></i> Login</a></li>
                     {% else %}
-                    <li><a href="{{ url_for('user', username=current_user.username) }}"><i class="fas fa-user"></i> Profile</a></li>
-                    <li><a href="{{ url_for('logout') }}"><i class="fas fa-sign-out-alt"></i> Logout</a></li>
+                    <li><a href="{{ url_for('main.user', username=current_user.username) }}"><i class="fas fa-user"></i> Profile</a></li>
+                    <li><a href="{{ url_for('auth.logout') }}"><i class="fas fa-sign-out-alt"></i> Logout</a></li>
                     {% endif %}
                 </ul>
             </div>
         </div>
         </nav>
 {% endblock %}
-        
 {% block content %}
     <div class="container">
         {% with messages = get_flashed_messages() %}
