+++ app/templates/email/reset_password.txt
@@ -1,11 +1,10 @@
 Dear {{ user.username }},
 
 To reset your password click on the following link:
-{{ url_for('reset_password', token=token, _external=True) }}
+{{ url_for('auth.reset_password', token=token, _external=True) }}
 
 If you have not requested a password reset simply ignore this message.
 
 
 Sincerely,
-
 The Spjald Team
