+++ app/templates/errors/404.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+
+{% block app_content %}
+    <h3>Not found</h3>
+    <p><a href="{{ url_for('main.index') }}">Back</a></p>
+{% endblock %}
