+++ app/templates/errors/500.html
@@ -3,5 +3,5 @@
 {% block app_content %}
     <h3>An unexpected error has occurred</h3>
     <p>The administrator has been notified. Sorry for the inconvenience!</p>
-    <p><a href="{{ url_for('index') }}">Back</a></p>
+    <p><a href="{{ url_for('main.index') }}">Back</a></p>
 {% endblock %}
