+++ app/templates/index.html
@@ -1,7 +1,5 @@
 {% extends "base.html" %}
-
 {% import 'bootstrap/wtf.html' as wtf %}
-
 {% block app_content %}
     <h3>Hi, {{ current_user.name }}!</h3>
         {% for post in posts %}
