+++ app/templates/user.html
@@ -1,5 +1,4 @@
 {% extends "base.html" %}
-
 {% block app_content %}
     <table class="table table-hover">
         <tr>
@@ -12,9 +11,9 @@
                 </p>
                 <p>{{ user.followers.count() }} followers, {{ user.followed.count() }} following.</p>
                 {% if not current_user.is_following(user) %}
-                <p><i class="fas fa-plus fa-sm"></i> <a href="{{ url_for('follow', username=user.username) }}"><b>Follow</b></a></p>
+                <p><i class="fas fa-plus fa-sm"></i> <a href="{{ url_for('main.follow', username=user.username) }}"><b>Follow</b></a></p>
                 {% else %}
-                <p><i class="fas fa-minus fa-sm"></i> <a href="{{ url_for('unfollow', username=user.username) }}"><b>Unfollow</b></a></p>
+                <p><i class="fas fa-minus fa-sm"></i> <a href="{{ url_for('main.unfollow', username=user.username) }}"><b>Unfollow</b></a></p>
                 {% endif %}
             </td>
         </tr>
