+++ config.py
@@ -1,7 +1,9 @@
 import os
+from dotenv import load_dotenv
 
 
 basedir = os.path.abspath(os.path.dirname(__file__))
+load_dotenv(os.path.join(basedir, '.env'))
 
 
 class Config(object):
