+++ requirements.txt
@@ -19,6 +19,7 @@ Jinja2==2.10.3
 Mako==1.1.0
 MarkupSafe==1.1.1
 mccabe==0.6.1
+pkg-resources==0.0.0
 pycodestyle==2.5.0
 pyflakes==2.1.1
 PyJWT==1.7.1
@@ -31,4 +32,5 @@ urllib3==1.25.7
 visitor==0.1.3
 webencodings==0.5.1
 Werkzeug==0.16.0
+wtf-tinymce==0.1.3
 WTForms==2.2.1
